+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add option to ignore IaC issues
 ### Fixed
 - folder-specific configs are availabe on opening projects, not only on restart of the IDE
+- display open source issues in Rider. Previously, as the project.assets.json is in a derived folder, it was filtered.
 ## [2.10.0]
 ### Changed
 - save git folder config in settings